restructure and add spdy to browser tests
