+++ package.json
@@ -37,18 +37,20 @@
     "node": "^4.3.0"
   },
   "devDependencies": {
+    "aegir": "^3.0.4",
     "bl": "^1.1.2",
     "buffer-loader": "0.0.1",
     "chai": "^3.5.0",
-    "aegir": "^3.0.4",
     "gulp": "^3.9.1",
     "istanbul": "^0.4.3",
     "libp2p-multiplex": "^0.2.1",
-    "libp2p-spdy": "^0.5.0",
+    "libp2p-spdy": "^0.6.0",
     "libp2p-tcp": "^0.5.1",
+    "libp2p-webrtc-star": "^0.1.0",
     "libp2p-websockets": "^0.5.0",
     "pre-commit": "^1.1.2",
-    "stream-pair": "^1.0.3"
+    "stream-pair": "^1.0.3",
+    "webrtcsupport": "^2.2.0"
   },
   "dependencies": {
     "babel-runtime": "^6.6.1",
