+++ test/browser.js
@@ -2,11 +2,7 @@
 'use strict'
 
 const expect = require('chai').expect
-const multiaddr = require('multiaddr')
-const Id = require('peer-id')
-const Peer = require('peer-info')
-const WebSockets = require('libp2p-websockets')
-const bl = require('bl')
+const w = require('webrtcsupport')
 
 const Swarm = require('../src')
 
@@ -17,102 +13,15 @@ describe('basics', () => {
   })
 })
 
-describe('transport - websockets', function () {
-  this.timeout(10000)
+require('./browser-00-transport-websockets.js')
 
-  var swarm
+if (w.support) {
+  require('./browser-01-transport-webrtc-star.js')
+}
 
-  before(() => {
-    const b58IdSrc = 'QmYzgdesgjdvD3okTPGZT9NPmh1BuH5FfTVNKjsvaAprhb'
-    // use a pre generated Id to save time
-    const idSrc = Id.createFromB58String(b58IdSrc)
-    const peerSrc = new Peer(idSrc)
-    swarm = new Swarm(peerSrc)
-  })
+require('./browser-02-swarm-with-muxing-plus-websockets.js')
 
-  it('add', (done) => {
-    swarm.transport.add('ws', new WebSockets(), () => {
-      expect(Object.keys(swarm.transports).length).to.equal(1)
-      done()
-    })
-  })
-
-  it('dial', (done) => {
-    const ma = multiaddr('/ip4/127.0.0.1/tcp/9100/ws')
-
-    const conn = swarm.transport.dial('ws', ma, (err, conn) => {
-      expect(err).to.not.exist
-    })
-
-    conn.pipe(bl((err, data) => {
-      expect(err).to.not.exist
-      expect(data.toString()).to.equal('hey')
-      done()
-    }))
-    conn.write('hey')
-    conn.end()
-  })
-})
-
-describe('high level API - 1st without stream multiplexing (on websockets)', function () {
-  this.timeout(10000)
-
-  var swarm
-  var peerDst
-
-  before(() => {
-    const b58IdSrc = 'QmYzgdesgjdvD3okTPGZT9NPmh1BuH5FfTVNKjsvaAprhb'
-    // use a pre generated Id to save time
-    const idSrc = Id.createFromB58String(b58IdSrc)
-    const peerSrc = new Peer(idSrc)
-    swarm = new Swarm(peerSrc)
-  })
-
-  after((done) => {
-    swarm.close(done)
-  })
-
-  it('add ws', () => {
-    swarm.transport.add('ws', new WebSockets())
-    expect(Object.keys(swarm.transports).length).to.equal(1)
-  })
-
-  it('create Dst peer info', () => {
-    const b58IdDst = 'QmYzgdesgjdvD3okTPGZT9NPmh1BuH5FfTVNKjsvaAprhb'
-    // use a pre generated Id to save time
-    const idDst = Id.createFromB58String(b58IdDst)
-    peerDst = new Peer(idDst)
-
-    const ma = multiaddr('/ip4/127.0.0.1/tcp/9200/ws')
-    peerDst.multiaddr.add(ma)
-  })
-
-  it('dial on protocol', (done) => {
-    swarm.dial(peerDst, '/echo/1.0.0', (err, conn) => {
-      expect(err).to.not.exist
-      conn.pipe(bl((err, data) => {
-        expect(err).to.not.exist
-        expect(data.toString()).to.equal('hey')
-        done()
-      }))
-      conn.write('hey')
-      conn.end()
-    })
-  })
-
-  it('dial to warm a conn', (done) => {
-    swarm.dial(peerDst, (err) => {
-      expect(err).to.not.exist
-      done()
-    })
-  })
-
-  it('dial on protocol, reuse warmed conn', (done) => {
-    swarm.dial(peerDst, '/echo/1.0.0', (err, conn) => {
-      expect(err).to.not.exist
-      conn.end()
-      conn.on('data', () => {}) // let it flow.. let it flooooow
-      conn.on('end', done)
-    })
-  })
-})
+if (w.support) {
+  require('./browser-03-swarm-with-muxing-plus-webrtc-star.js')
+  require('./browser-04-swarm-with-muxing-plus-websockets-and-webrtc-star.js')
+}
